Split bookmarklet into browser and server
The browser bookmarklet is for capturing resource timing data when we
cannot connect to the capture server (to load the main client.js file and
to POST the reults). This bookmarklet outputs the results to a TEXTAREA
created in the browser page.

The server bookmarklet does not contain the TEXTAREA code and relies on
being able to connect to the capture server.

Added npm scripts to watch and build all the JS.

Update README for client build.